[BREAKING][rollout] refactor: remove the legacy Harbor backend
Delete the pre-v0.3 `HarborBackend` and its container-side half. It is
superseded by `HarborBackendV2`, which packages user code into a wheel
instead of mounting the SDK into the task environment, so task images
stay pure task environments.

Removed:
  - `harbor/backend.py` (`HarborBackend`)
  - `harbor/agent_adapter.py`, `harbor/agent_runner.py`,
    `harbor/grader_runner.py` — the installed-agent + verifier pair only
    that backend ever invoked
  - `OsmosisInstalledAgent` and `HarborAgentWorkflowContext` public names.
    The context was already inert: the in-container runner built a plain
    `AgentWorkflowContext`, so its `environment` field was always None.

The deleted module also hosted plumbing `HarborBackendV2` depends on, so
that moves to modules neither backend owns:
  - `harbor/trial.py` — trial naming, `PendingTrial`, hook-event readers
  - `harbor/environment.py` — SkyPilot placement, docker-runtime helpers
  - `ensure_import_path` joins `to_import_path` in `rollout/utils/imports.py`

Without that split `artifacts.py -> backend.py -> artifacts.py` becomes a
circular import the moment the surviving backend takes the `backend.py`
name (next commit).

Also annotates `native_agent_kwargs`, which `--verifytypes` flagged as
inferrable-differently.
