+++ osmosis_ai/rollout/backend/harbor/__init__.py
@@ -1,11 +1,7 @@
-from osmosis_ai.rollout.backend.harbor.agent_adapter import OsmosisInstalledAgent
-from osmosis_ai.rollout.backend.harbor.backend import HarborBackend
 from osmosis_ai.rollout.backend.harbor.backend_v2 import HarborBackendV2
 from osmosis_ai.rollout.backend.harbor.tasks import TaskMode
 
 __all__ = [
-    "HarborBackend",
     "HarborBackendV2",
-    "OsmosisInstalledAgent",
     "TaskMode",
 ]
